+++ World.h
@@ -14,7 +14,7 @@
 #include <string>
 
 typedef cv::Vec3b Color;
-typedef cv::Vec3f Vec;
+typedef cv::Vec3d Vec;
 
 class World {
 private:
@@ -22,24 +22,24 @@ private:
     std::vector<Light*> lightSources;
     std::string name;
     Color bgColor;
-    float aIntensity;
+    double aIntensity;
 
-    Color rayTracing(Ray& ray, int depth=7, float epsilon = .0) const;
-    Object* hit(float& t, Ray& ray, float epsilon = .0, double max = INT_MAX) const;
+    Color rayTracing(Ray& ray, int depth=7, double epsilon = .0) const;
+    Object* hit(double& t, Ray& ray, double epsilon = .0, double max = INT_MAX) const;
     static Vec normalize(Vec v);
     static void printVec(Vec& v);
     static void printColor(Color& c);
-    bool refract(Vec& d, Vec&n, float nt, Vec& t) const;
-    void render(float l, float r, float b, float t, float d, int nx, int ny, Camera& cam, cv::Mat& mat) const;
+    bool refract(Vec& d, Vec&n, double nt, Vec& t) const;
+    void render(double l, double r, double b, double t, double d, int nx, int ny, Camera& cam, cv::Mat& mat) const;
     Color superSample(int i, int j, int nx, int ny, cv::Mat &image) const;
 
 public:
-    World(Color bgColor = Color(255,255,255), float aIntensity = 0, std::string name="Brave New World");
+    World(Color bgColor = Color(255,255,255), double aIntensity = 0, std::string name="Brave New World");
     std::string getName() const;
     void setName(std::string name);
     void addObject(Object* obj);
     void addLightSource(Light* l);
-    void render(float l, float r, float b, float t, float d, int nx, int ny,
+    void render(double l, double r, double b, double t, double d, int nx, int ny,
                 Camera& cam, bool ssaa=false) const;
     ~World();
 };
